feat: router and routes in new folder to alocate private route component
